requirements.txt
scrapy
cohere
langchain
langchain-cohere
python-dotenv
pyrate-limiter
tenacity